captcha-solvers/requirements.txt
# CF Clearance Scraper - 统一Python依赖
# 包含所有验证码解决器的依赖包

# ===========================================
# 通用依赖 (Common Dependencies)
# ===========================================
python-dotenv>=1.0.0
pathlib2; python_version < '3.4'

# ===========================================
# hCAPTCHA 解决器依赖 (hCAPTCHA Solver)
# ===========================================
# AI 和机器学习
hcaptcha-challenger>=0.18.0
google-genai>=1.19.0,<2.0.0

# 浏览器自动化
playwright>=1.40.0

# 图像处理
opencv-python>=4.11.0.86,<5.0
numpy>=2.0,<3.0
pillow>=11.1.0
matplotlib>=3.10.3,<4.0.0

# 网络和数据处理
httpx[http2]>=0.25.0
msgpack>=1.1.0,<2.0.0
tenacity>=9.1.2

# 配置和日志
pydantic-settings>=2.0.0
loguru>=0.7.0
typer>=0.15.4
pytz

# ===========================================
# reCAPTCHA 解决器依赖 (reCAPTCHA Solver)
# ===========================================
# 浏览器自动化 (不同框架)
DrissionPage>=4.0.0

# 音频处理
pydub>=0.25.0
SpeechRecognition>=3.10.0

# ===========================================
# 可选依赖 (Optional Dependencies)
# ===========================================
# 音频格式支持 (for pydub)
# 取消注释以下行如果需要更多音频格式支持：
# ffmpeg-python>=0.2.0  # 需要系统安装 ffmpeg
# simpleaudio>=1.0.4    # 音频播放支持

# 语音识别引擎扩展 (for SpeechRecognition)
# 取消注释以下行如果需要离线语音识别：
# vosk>=0.3.45          # 离线语音识别
# whisper>=1.0          # OpenAI Whisper 模型

# ===========================================
# 系统兼容性说明
# ===========================================
# Windows 用户注意事项：
# 1. 安装 Visual C++ Redistributable
# 2. 某些包可能需要 Microsoft Visual C++ 14.0+
#
# macOS 用户注意事项：
# 1. 可能需要安装 Xcode Command Line Tools
# 2. 某些包需要 Homebrew 安装的依赖
#
# Linux 用户注意事项：
# 1. 需要安装 python3-dev, python3-pip
# 2. 图像处理包可能需要额外的系统库
#
# ===========================================
# 安装命令示例
# ===========================================
# 标准安装：
# pip install -r requirements.txt
#
# 用户安装（推荐）：
# pip install --user -r requirements.txt
#
# 升级已有包：
# pip install --upgrade -r requirements.txt
#
# 强制重新安装：
# pip install --force-reinstall -r requirements.txt